+++ src/idt.c
@@ -21,17 +21,17 @@ void int0h_handler() { print("devide by zero error"); }
 void idtSet(int interuptNumber, void *address) {
 
   idtDesc *desc = &intDescriptors[interuptNumber];
-  desc->offset_1 = (unsigned long)address & 0x0000ffff;
+  desc->offset_1 = (uint32_t)address & 0x0000ffff;
   desc->selector = codeSelector;
   desc->zero = 0x00;
   desc->typeAttr = 0xEE;
-  desc->offset_2 = (unsigned long)address >> 16;
+  desc->offset_2 = (uint32_t)address >> 16;
 }
 void initializeIdt() {
   memset(intDescriptors, 0, sizeof(intDescriptors) / sizeof(intDescriptors[0]));
 
   idtrDescriptor.limit = sizeof(intDescriptors) - 1;
-  idtrDescriptor.base = (unsigned long)intDescriptors;
+  idtrDescriptor.base = (uint32_t)intDescriptors;
   for (int i = 0; i < 512; i++) {
     idtSet(i, no_interrupt);
   }
